+++ src/main/java/community/flock/dialogflow/ci/json/Capability.java
@@ -1,0 +1,13 @@
+package community.flock.dialogflow.ci.json;
+
+public class Capability {
+	private String name;
+
+	public String getName() {
+		return name;
+	}
+
+	public void setName(String name) {
+		this.name = name;
+	}
+}