+++ src/main/java/community/flock/dialogflow/ci/json/Surface.java
@@ -1,0 +1,15 @@
+package community.flock.dialogflow.ci.json;
+
+import java.util.List;
+
+public class Surface {
+	private List<Capability> capabilites;
+
+	public List<Capability> getCapabilites() {
+		return capabilites;
+	}
+
+	public void setCapabilites(List<Capability> capabilites) {
+		this.capabilites = capabilites;
+	}
+}