+++ src/main/java/community/flock/dialogflow/ci/json/User.java
@@ -1,0 +1,35 @@
+package community.flock.dialogflow.ci.json;
+
+import java.util.List;
+
+public class User {
+	private String lastSeen;
+	private String locale;
+	private String userId;
+	private List<String> permissions;
+	
+	public String getLastSeen() {
+		return lastSeen;
+	}
+	public void setLastSeen(String lastSeen) {
+		this.lastSeen = lastSeen;
+	}
+	public String getLocale() {
+		return locale;
+	}
+	public void setLocale(String locale) {
+		this.locale = locale;
+	}
+	public String getUserId() {
+		return userId;
+	}
+	public void setUserId(String userId) {
+		this.userId = userId;
+	}
+	public List<String> getPermissions() {
+		return permissions;
+	}
+	public void setPermissions(List<String> permissions) {
+		this.permissions = permissions;
+	}
+}