+++ RELEASES.md
@@ -1,3 +1,10 @@
+## RELEASE_4
+
+* Can give a device (phone, speaker, smart display) at the start of scenarios
+* Can give a user at the start (its userId)
+! No longer accepts the name of the Google Home applicaton in the first sentence,
+    step is just: "... met de Google Home applicatie"
+
 ## RELEASE_3
 
 * Prints the intents which were covered and uncovered during the test run
@@ -7,7 +14,6 @@
 * Add support for basic cards
 * Also search in simple responses for speech
 * Support all combinations of "she" (ze) and "the assistant" (de assistent)
-
 
 ## RELEASE_1